MET-datacenter-wind-analysis/requirements.txt
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.4
Pillow==9.4.0
streamlit==1.16.0
